+++ src/App.tsx
@@ -4,6 +4,7 @@ import Layout from './components/layout/Layout'
 import ClientesPage from './pages/clientes'
 import ArticulosPage from './pages/articulos'
 import ProveedoresPage from './pages/proveedores'
+import ComprasPage from './pages/compras'
 import FacturacionPage from './pages/facturacion'
 import LoginPage from './pages/login'
 import UsersPage from './pages/users'
@@ -90,6 +91,7 @@ function AppRoutes() {
           <Route path="clientes" element={<ClientesPage />} />
           <Route path="articulos" element={<ArticulosPage />} />
           <Route path="proveedores" element={<ProveedoresPage />} />
+          <Route path="compras" element={<ComprasPage />} />
           <Route path="facturacion" element={<FacturacionPage />} />
           <Route path="pedidos" element={<PedidosPage />} />
           <Route path="users" element={<UsersPage />} />